Add guidelines-based argument for correctness of paxos.
Rather than giving a particular implementation, we give some facts
that are (a) relatively feasible to reason about and (b) sufficient to
guarantee the safety of the implementation.

The benefit of this is that it puts very few constraints on the
implementation. In particular, the student can choose what to do in
heartbeats, what to do with their timers, etc. -- as long as they meet
the requirements, then their protocol will be safe.