Show that only one value will be chosen in a full execution.
This change introduces the concept of a state being "reachable from"
another state. The relevance is that for two states to be "in the same
execution", one must be reachable (from an initial state) and the
other must be reachable from the first.

Under the assumption that `voted` is a stable predicate, we show that
if a state `dst` is reachable from a reachable state `src`, and values
are chosen at `src` and `dst`, then both chosen values must be the
same. The argument is that the stability of `voted` implies the
stability of `chosen`, so if a value is chosen at `src`, it is also
chosen at `dst`. The first safety theorem shows that both values
are the same.